refactor(usuarios): modularize table and state
