+++ zen-cat-frontend/ZenCat-Front-Admin/src/routes/usuarios/index.tsx
@@ -1,6 +1,6 @@
-import React, { useState, useMemo } from "react";
+import React, { useState } from "react";
 import { createFileRoute, useNavigate, Link } from "@tanstack/react-router";
-import { Search, Filter, Download, ArrowUpDown, MoreHorizontal, Plus, Trash, Loader2 } from "lucide-react";
+import { Plus, Loader2 } from "lucide-react";
 import HeaderDescriptor from '@/components/common/header-descriptor';
 import { Button } from "@/components/ui/button";
 import { User } from "@/types/user";
@@ -19,22 +19,7 @@ import {
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { usuariosApi } from '@/api/usuarios/usuarios';
 import { toast } from "sonner";
-import {
-  ColumnDef,
-  useReactTable,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  getFilteredRowModel,
-  SortingState,
-  ColumnFiltersState,
-  VisibilityState,
-  PaginationState,
-} from '@tanstack/react-table';
-import { Checkbox } from '@/components/ui/checkbox';
-import { DataTable } from '@/components/common/data-table/data-table';
-import { DataTableToolbar } from '@/components/common/data-table/data-table-toolbar';
-import { DataTablePagination } from '@/components/common/data-table/data-table-pagination';
+import { UsersTable } from '@/components/users/table';
 
 export const Route = createFileRoute('/usuarios/')({
   component: UsuariosComponent,
@@ -43,12 +28,6 @@ export const Route = createFileRoute('/usuarios/')({
 function UsuariosComponent() {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-  const [sorting, setSorting] = useState<SortingState>([]);
-  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
-  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
-  const [rowSelection, setRowSelection] = useState({});
-  const [globalFilter, setGlobalFilter] = useState('');
-  const [pagination, setPagination] = useState<PaginationState>({ pageIndex: 0, pageSize: 10 });
 
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const [userToDelete, setUserToDelete] = useState<User | null>(null);
@@ -65,132 +44,23 @@ function UsuariosComponent() {
     onSuccess: (_, id) => {
       toast.success('Usuario eliminado', { description: `Usuario eliminado exitosamente` });
       queryClient.invalidateQueries({ queryKey: ['usuarios'] });
-      setRowSelection({});
     },
     onError: (err) => {
       toast.error('Error al eliminar', { description: err.message });
     },
   });
 
-  const columns = useMemo<ColumnDef<User>[]>(() => [
-    {
-      id: 'select',
-      header: ({ table }) => (
-        <Checkbox
-          checked={table.getIsAllPageRowsSelected() || (table.getIsSomePageRowsSelected() && 'indeterminate')}
-          onCheckedChange={(v) => table.toggleAllPageRowsSelected(!!v)}
-          aria-label="Select all"
-        />
-      ),
-      cell: ({ row }) => (
-        <Checkbox
-          checked={row.getIsSelected()}
-          onCheckedChange={(v) => row.toggleSelected(!!v)}
-          aria-label="Select row"
-        />
-      ),
-      enableSorting: false,
-      enableHiding: false,
-    },
-    {
-      accessorKey: 'name',
-      header: ({ column }) => (
-        <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}>
-          Nombres <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-      cell: ({ row }) => <div>{row.getValue('name')}</div>,
-    },
-    {
-      accessorKey: 'address',
-      header: 'Dirección',
-      cell: ({ row }) => <div>{row.getValue('address') || '-'}</div>,
-    },
-    {
-      accessorKey: 'district',
-      header: 'Distrito',
-      cell: ({ row }) => <div>{row.getValue('district') || '-'}</div>,
-    },
-    {
-      accessorKey: 'phone',
-      header: 'Teléfono',
-      cell: ({ row }) => <div>{row.getValue('phone') || '-'}</div>,
-    },
-    {
-      accessorKey: 'email',
-      header: ({ column }) => (
-        <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}>
-          Email <ArrowUpDown className="ml-2 h-4 w-4" />
-        </Button>
-      ),
-    },
-    {
-      id: 'memberships',
-      header: 'Membresías',
-      cell: ({ row }) => {
-        const user = row.original;
-        return (
-          <Button
-            className="h-9 px-4 bg-white border border-neutral-300 text-black rounded-lg flex items-center gap-2 shadow-sm hover:bg-black hover:text-white hover:border-black hover:shadow-md transition-all duration-200"
-            onClick={() => navigate({ to: '/usuarios/ver_membresia', search: { id: user.id } })}
-          >
-            Membresías ...
-          </Button>
-        );
-      },
-    },
-    {
-      id: 'actions',
-      cell: ({ row }) => {
-        const user = row.original;
-        return (
-          <div className="flex items-center space-x-2">
-            <Button
-              className="h-8 w-8 p-0 bg-white text-black border border-black rounded-full flex items-center justify-center hover:bg-red-100 hover:shadow-md transition-all duration-200"
-              onClick={(e) => {
-                e.stopPropagation();
-                navigate({ 
-                  to: '/usuarios/editar',
-                  search: { id: user.id },
-                  replace: true
-                });
-              }}
-            >
-              <MoreHorizontal className="h-4 w-4" />
-            </Button>
-            <Button
-              className="h-8 w-8 p-0 bg-white text-black border border-black rounded-full flex items-center justify-center hover:bg-red-100 hover:shadow-md transition-all duration-200"
-              onClick={(e) => {
-                e.stopPropagation();
-                setUserToDelete(user);
-                setIsDeleteModalOpen(true);
-              }}
-            >
-              <Trash className="h-4 w-4" />
-            </Button>
-          </div>
-        );
-      },
-    },
-  ], [navigate]);
+  const handleEdit = (user: User) => {
+    navigate({
+      to: '/usuarios/editar',
+      search: { id: user.id },
+      replace: true,
+    });
+  };
 
-  const table = useReactTable({
-    data: usersData || [],
-    columns,
-    state: { sorting, columnFilters, columnVisibility, rowSelection, globalFilter, pagination },
-    onSortingChange: setSorting,
-    onColumnFiltersChange: setColumnFilters,
-    onGlobalFilterChange: setGlobalFilter,
-    onPaginationChange: setPagination,
-    onColumnVisibilityChange: setColumnVisibility,
-    onRowSelectionChange: setRowSelection,
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    manualPagination: false,
-    enableRowSelection: true,
-  });
+  const handleViewMemberships = (user: User) => {
+    navigate({ to: '/usuarios/ver_membresia', search: { id: user.id } });
+  };
 
   const btnSizeClasses = "h-11 w-28 px-4";
 
@@ -245,21 +115,15 @@ function UsuariosComponent() {
           <Loader2 className="h-16 w-16 animate-spin text-gray-500" />
         </div>
       ) : (
-        <div className="-mx-4 flex-1 overflow-auto px-4 py-2">
-          <DataTableToolbar
-            table={table}
-            filterPlaceholder="Buscar usuarios..."
-            showSortButton
-            showFilterButton
-            showExportButton
-            onFilterClick={() => console.log('Filtrar')}
-            onExportClick={() => console.log('Exportar')}
-          />
-          <div className="flex-1 overflow-hidden rounded-md border">
-            <DataTable table={table} columns={columns} />
-          </div>
-          <DataTablePagination table={table} />
-        </div>
+        <UsersTable
+          data={usersData || []}
+          onEdit={handleEdit}
+          onDelete={(u) => {
+            setUserToDelete(u);
+            setIsDeleteModalOpen(true);
+          }}
+          onViewMemberships={handleViewMemberships}
+        />
       )}
 
       <AlertDialog open={isDeleteModalOpen} onOpenChange={setIsDeleteModalOpen}>
